Fix regression introduced by 78fb5c7
The commit introduced default values for configuration parameters that
can be specified on the command line, but it set those default values
too early.
This led to it overwriting those values when they are specified in the
XML file.
This change changes it so that the absence of a command line parameter
leads to it being set as null, rather than the default value.
The default value is set during the last stage of the Ganglia
configuration, which is after the XML file has been read.