attempt at #3, pattern in MBean ObjectName
Each mbean name given in the XML configuration file is expanded using
the query method of MBeanServer.

The query method takes in an ObjectName and returns all registered MBeans that
matches it.

If there is an error, for example a typo in the configuration, the malformed
MBean isn't added to the MBeanSampler.